+++ src/index.ts
@@ -3,3 +3,4 @@ export * from "./collection";
 export * from "./pipeAsync";
 export * from "./result";
 export * from "./sync";
+export * from "./utils";