+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.5]
+
+### Added
+
+- parseJson and attempt utils
+
+### Fixed
+
+- Incorrect dependency on rollup
+
 ## [1.2.4]
 
 ### Fixed